Move struct to proper files
